Renaming mei export file and removing old version.
