+++ partitura/io/__init__.py
@@ -15,7 +15,7 @@ from .importkern import load_kern
 from .importparangonada import load_parangonada_csv
 from .exportparangonada import save_parangonada_csv
 from .importmusic21 import load_music21
-from .mei_export_v2 import save_mei
+from .exportmei import save_mei
 from partitura.utils.misc import (
     deprecated_alias,
     deprecated_parameter,
